Try to deal better with merged cells
Cache where they are and at the same time create a couple of lookup
tables so that we can go easily from row/column indices to the
Excel cell in question.  There are two versions of this for now; one
with the merged cells and one without.  A mask might be a better way
of doing this, though.