Detect multiple tables in a worksheet
Uses the flood fill algorithm to identify contiguous regions of cells,
with the constraint that they should be rectangular.  The `split_tables`
function (not exported) will detect all regions and then split the
worksheet into a list of tables.  This will probably be the first
step in rationalising a suitably complicated worksheet.